移动文件
[Chore]
文件夹header重命名为setting，部分改写为tsx。
