+++ src/renderer/src/components/search/Search.vue
@@ -45,7 +45,7 @@ import Header from './Header.vue'
 import Footer from './Footer.vue'
 import Task from '@/components/task/Task.vue'
 import { useTasks } from '@/hooks/useTasks'
-import CreatorInSearch from '@/components/CreatorInSearch.vue'
+import CreatorInSearch from '@/components/search/CreatorInSearch.vue'
 import { Search, SearchType } from '@/assets/js/search'
 import { Items } from '@/hooks/useList'
