+++ src/renderer/src/components/setting/lib/Avatar.tsx
