+++ src/renderer/src/components/setting/lib/ThemeColors.tsx
@@ -1,0 +1,31 @@
+import useTheme from '@/hooks/useTheme'
+
+const theme = useTheme()
+
+const ThemeColors = () => (
+    <ul class="flex flex-col gap-2">
+        {
+            theme.getThemeColor().map((e, index) => (
+                <li
+                    key={index}
+                    class="flex items-center justify-start"
+                >
+                    <label>
+                        <p>{{ e }}</p>
+                        <md-radio
+                            name="theme"
+                            class="absolute right-0"
+                            onClick={() => theme.set({
+                                color: e
+                            })}
+                        ></md-radio>
+                    </label>
+                </li>
+            ))
+        }
+    </ul>
+)
+
+export {
+    ThemeColors
+}
